# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ROMP API",
    "version": "1.0.0",
    "description": "The ROMP commerce API. Every write in the system passes through here. Generated from Zod contracts."
  },
  "paths": {
...
    "/v1/admin/orders": {
      "get": {
        "summary": "List orders — newest first, filterable by status or fulfilment, or searched by humanId",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/OrderStatus"
            }
          },
          {
            "name": "fulfilmentStatus",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/FulfilmentStatus"
            }
          },
          {
            "name": "humanId",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/HumanOrderId"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of orders, newest first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderListResponse"
                }
              }
            }
          },
          "400": {
            "description": "`VALIDATION_FAILED` (Validation failed) — https://romp.dev/errors/validation-failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHENTICATED` (Authentication required) — https://romp.dev/errors/unauthenticated",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "`FORBIDDEN` (Not permitted) — https://romp.dev/errors/forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/orders/{id}/verify-payment": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "post": {
        "summary": "Verify a payment: commit stock and mark the order paid (exact amount)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyPaymentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The paid order.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderView"
                }
              }
            }
          },
          "400": {
            "description": "`VALIDATION_FAILED` (Validation failed) — https://romp.dev/errors/validation-failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHENTICATED` (Authentication required) — https://romp.dev/errors/unauthenticated",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "`FORBIDDEN` (Not permitted) — https://romp.dev/errors/forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "`NOT_FOUND` (Not found) — https://romp.dev/errors/not-found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "`PAYMENT_AMOUNT_MISMATCH` (Payment amount does not match the order total) — https://romp.dev/errors/payment-amount-mismatch; `INVALID_STATE_TRANSITION` (Invalid state transition) — https://romp.dev/errors/invalid-state-transition",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "`RATE_LIMITED` (Too many requests) — https://romp.dev/errors/rate-limited",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/orders/{id}/reject-payment": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "post": {
        "summary": "Reject a payment the admin could not match; the customer may resubmit",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectPaymentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The rejected order, with the reason.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderView"
                }
              }
            }
          },
          "400": {
            "description": "`VALIDATION_FAILED` (Validation failed) — https://romp.dev/errors/validation-failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHENTICATED` (Authentication required) — https://romp.dev/errors/unauthenticated",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "`FORBIDDEN` (Not permitted) — https://romp.dev/errors/forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "`NOT_FOUND` (Not found) — https://romp.dev/errors/not-found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "`INVALID_STATE_TRANSITION` (Invalid state transition) — https://romp.dev/errors/invalid-state-transition",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "`RATE_LIMITED` (Too many requests) — https://romp.dev/errors/rate-limited",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/orders/{id}/fulfilment": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "post": {
        "summary": "Advance fulfilment: pack, ship (with carrier and tracking), deliver, or hold",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FulfilmentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The order at its new fulfilment stage.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderView"
                }
              }
            }
          },
          "400": {
            "description": "`VALIDATION_FAILED` (Validation failed) — https://romp.dev/errors/validation-failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHENTICATED` (Authentication required) — https://romp.dev/errors/unauthenticated",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "`FORBIDDEN` (Not permitted) — https://romp.dev/errors/forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "`NOT_FOUND` (Not found) — https://romp.dev/errors/not-found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "`INVALID_STATE_TRANSITION` (Invalid state transition) — https://romp.dev/errors/invalid-state-transition",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "`RATE_LIMITED` (Too many requests) — https://romp.dev/errors/rate-limited",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/orders/{id}/cancel": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "post": {
        "summary": "Cancel an order: release a held reservation, or restock a paid order",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelOrderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The cancelled order.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderView"
                }
              }
            }
          },
          "400": {
            "description": "`VALIDATION_FAILED` (Validation failed) — https://romp.dev/errors/validation-failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHENTICATED` (Authentication required) — https://romp.dev/errors/unauthenticated",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "`FORBIDDEN` (Not permitted) — https://romp.dev/errors/forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "`NOT_FOUND` (Not found) — https://romp.dev/errors/not-found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "`INVALID_STATE_TRANSITION` (Invalid state transition) — https://romp.dev/errors/invalid-state-transition",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "`RATE_LIMITED` (Too many requests) — https://romp.dev/errors/rate-limited",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/refunds": {
      "post": {
        "summary": "Issue a full or partial refund (owner-only, append-only, optional restock)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueRefundRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The refund result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueRefundResponse"
                }
              }
            }
          },
          "400": {
            "description": "`VALIDATION_FAILED` (Validation failed) — https://romp.dev/errors/validation-failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHENTICATED` (Authentication required) — https://romp.dev/errors/unauthenticated",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "`FORBIDDEN` (Not permitted) — https://romp.dev/errors/forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "`NOT_FOUND` (Not found) — https://romp.dev/errors/not-found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "`REFUND_EXCEEDS_REFUNDABLE` (Refund exceeds refundable amount) — https://romp.dev/errors/refund-exceeds-refundable; `INVALID_STATE_TRANSITION` (Invalid state transition) — https://romp.dev/errors/invalid-state-transition",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "`RATE_LIMITED` (Too many requests) — https://romp.dev/errors/rate-limited",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Money": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/Money",
        "type": "integer",
        "minimum": 0,
        "maximum": 9007199254740991
      },
...
      "Cursor": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/Cursor",
        "type": "string",
        "minLength": 1,
        "maxLength": 2048
      },
...
      "HumanOrderId": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/HumanOrderId",
        "type": "string",
        "pattern": "^[A-Z]{2,8}-\\d{4,12}$"
      },
...
      "OrderStatus": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/OrderStatus",
        "type": "string",
        "enum": [
          "awaiting_payment",
          "pending_verification",
          "paid",
          "payment_rejected",
          "expired",
          "cancelled",
          "refunded"
        ]
      },
      "FulfilmentStatus": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/FulfilmentStatus",
        "type": "string",
        "enum": [
          "unfulfilled",
          "packed",
          "shipped",
          "delivered",
          "on_hold",
          "cancelled"
        ]
      },
...
      "RefundMode": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/RefundMode",
        "type": "string",
        "enum": [
          "full",
          "partial"
        ]
      },
      "RefundReason": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/RefundReason",
        "type": "string",
        "enum": [
          "customer_cancelled",
          "out_of_stock",
          "damaged",
          "wrong_item",
          "delivery_failed",
          "overpayment",
          "adjustment",
          "duplicate_refund",
          "other"
        ]
      },
...
      "ErrorCode": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/ErrorCode",
        "type": "string",
        "enum": [
          "VALIDATION_FAILED",
          "FILTER_LIMIT_EXCEEDED",
          "UNSUPPORTED_QUERY",
          "UNAUTHENTICATED",
          "FORBIDDEN",
          "NOT_FOUND",
          "IDENTIFIER_TAKEN",
          "INSUFFICIENT_STOCK",
          "RESERVATION_EXPIRED",
          "DUPLICATE_PAYMENT_REFERENCE",
          "PAYMENT_AMOUNT_MISMATCH",
          "INVALID_STATE_TRANSITION",
          "VARIANT_IN_USE",
          "REFUND_EXCEEDS_REFUNDABLE",
          "WEAK_PASSWORD",
          "RATE_LIMITED",
          "INTERNAL"
        ]
      },
      "ValidationIssue": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/ValidationIssue",
        "type": "object",
        "properties": {
          "path": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "path",
          "message"
        ],
        "additionalProperties": false
      },
      "ProblemDetails": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/ProblemDetails",
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            }
          },
          "retryAfterSeconds": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "additionalProperties": false
      },
...
      "VerifyPaymentRequest": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/VerifyPaymentRequest",
        "type": "object",
        "properties": {
          "paidAmountMinor": {
            "$ref": "#/components/schemas/Money"
          }
        },
        "required": [
          "paidAmountMinor"
        ],
        "additionalProperties": false
      },
      "RejectPaymentRequest": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/RejectPaymentRequest",
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          }
        },
        "required": [
          "reason"
        ],
        "additionalProperties": false
      },
      "IssueRefundRequest": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/IssueRefundRequest",
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1500
          },
          "mode": {
            "$ref": "#/components/schemas/RefundMode"
          },
          "amountMinor": {
            "$ref": "#/components/schemas/Money"
          },
          "reason": {
            "$ref": "#/components/schemas/RefundReason"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ]
          },
          "outwardUpiRef": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ]
          },
          "restock": {
            "type": "boolean"
          }
        },
        "required": [
          "orderId",
          "mode",
          "amountMinor",
          "reason",
          "note",
          "outwardUpiRef",
          "restock"
        ],
        "additionalProperties": false
      },
      "IssueRefundResponse": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/IssueRefundResponse",
        "type": "object",
        "properties": {
          "refundId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1500
          },
          "orderId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1500
          },
          "status": {
            "type": "string",
            "enum": [
              "paid",
              "refunded"
            ]
          },
          "refundedMinor": {
            "$ref": "#/components/schemas/Money"
          }
        },
        "required": [
          "refundId",
          "orderId",
          "status",
          "refundedMinor"
        ],
        "additionalProperties": false
      },
...
      "AdminOrderListResponse": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/AdminOrderListResponse",
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderView"
            }
          },
          "nextCursor": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Cursor"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "items",
          "nextCursor"
        ],
        "additionalProperties": false
      },
      "FulfilmentRequest": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/FulfilmentRequest",
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/FulfilmentStatus"
          },
          "carrier": {
            "default": null,
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ]
          },
          "trackingNo": {
            "default": null,
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ]
          },
          "holdReason": {
            "default": null,
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "status",
          "carrier",
          "trackingNo",
          "holdReason"
        ],
        "additionalProperties": false
      },
      "CancelOrderRequest": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/CancelOrderRequest",
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "restock": {
            "default": false,
            "type": "boolean"
          }
        },
        "required": [
          "reason",
          "restock"
        ],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "firebaseIdToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Firebase ID token. Admin routes additionally require a role custom claim and verify token revocation on every request."
      }
    }
  }
}